pyproject.toml
[tool.poetry]
name = "vllm-deployment"
version = "0.1.0"
description = ""
authors = ["pandego"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.2.7"
langchain-community = "^0.2.9"
vllm = "^0.5.1"
openai = "^1.35.13"
ipykernel = "^6.29.5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
